main: move main/lcpp.c to parsers/meta-cpreprocessor.c
