+++ parsers/c.c
@@ -18,7 +18,7 @@
 
 #include "debug.h"
 #include "entry.h"
-#include "lcpp.h"
+#include "meta-cpreprocessor.h"
 #include "keyword.h"
 #include "options.h"
 #include "parse.h"
