+++ parsers/cxx/cxx_parser.c
@@ -18,7 +18,7 @@
 
 #include "parse.h"
 #include "vstring.h"
-#include "lcpp.h"
+#include "../meta-cpreprocessor.h"
 #include "debug.h"
 #include "keyword.h"
 #include "read.h"
