+++ parsers/cxx/cxx_parser_tokenizer.c
@@ -16,7 +16,7 @@
 
 #include "parse.h"
 #include "vstring.h"
-#include "lcpp.h"
+#include "../meta-cpreprocessor.h"
 #include "debug.h"
 #include "keyword.h"
 #include "read.h"
