+++ parsers/cxx/cxx_tag.c
@@ -14,7 +14,7 @@
 #include "cxx_parser_internal.h"
 
 #include "entry.h"
-#include "lcpp.h"
+#include "../meta-cpreprocessor.h"
 #include "routines.h"
 #include "xtag.h"
