+++ win32/ctags_vs2013.vcxproj
@@ -132,7 +132,7 @@
     <ClCompile Include="..\parsers\json.c" />
     <ClCompile Include="..\main\keyword.c" />
     <ClCompile Include="..\main\kind.c" />
-    <ClCompile Include="..\main\lcpp.c" />
+    <ClCompile Include="..\parsers\cpreprocessor.c" />
     <ClCompile Include="..\parsers\lisp.c" />
     <ClCompile Include="..\parsers\iniconf.c" />
     <ClCompile Include="..\main\lregex.c" />
@@ -237,7 +237,7 @@
     <ClInclude Include="..\main\htable.h" />
     <ClInclude Include="..\main\keyword.h" />
     <ClInclude Include="..\main\kind.h" />
-    <ClInclude Include="..\main\lcpp.h" />
+    <ClInclude Include="..\parsers\meta-cpreprocessor.h" />
     <ClInclude Include="..\main\main.h" />
     <ClInclude Include="..\main\nestlevel.h" />
     <ClInclude Include="..\main\options.h" />
